requirements.txt
streamlit>=1.32.0
streamlit-folium>=0.11.0
pyyaml>=6.0
pandas>=2.0.0
folium>=0.14.0
requests>=2.28.0
